# Conflicts:
#	.idea/misc.xml
#	vigames/db.sqlite3

+++ vigames/games/views.py
@@ -1,23 +1,20 @@
 from datetime import timedelta, date
-
-from django.http import Http404
+from django.http import Http404, HttpResponse
 from rest_framework import status
 from rest_framework.generics import (RetrieveUpdateDestroyAPIView, ListAPIView)
 from rest_framework.permissions import IsAuthenticated
-from .models import Game, Rating, Account, Posts
 from rest_framework.response import Response
 from rest_framework.views import APIView
-
-from .models import Account, Posts
+from .models import Account, Posts, Game, Rating, Category, FAQ, Comments_Post, Comments_Game
 from .permissions import IsOwnerProfileOrReadOnly
-from .serializers import AccountSerializer, OutputAllNews, GameSerializer, OutputPost, RatingSerializer, \
-    CommentsNewsSerializer, PostSerializer
+from .serializers import AccountSerializer, OutputAllNews, GameSerializer, OutputPost, \
+    RatingSerializer, CommentsNewsSerializer, PostSerializer, FaqSerializer, CommentsGameSerializer, OrderSerializer
+from django.contrib.auth.models import User
 
-
-class UserProfileDetailView(RetrieveUpdateDestroyAPIView):
-    queryset = Account.objects.all()
-    serializer_class = AccountSerializer
-    permission_classes = [IsOwnerProfileOrReadOnly, IsAuthenticated]
+#class UserProfileDetailView(RetrieveUpdateDestroyAPIView):
+    #queryset = Account.objects.all()
+    #serializer_class = AccountSerializer
+    #permission_classes = [IsOwnerProfileOrReadOnly, IsAuthenticated]
 
 
 class OutputAllNewsView(APIView):
@@ -27,6 +24,37 @@ class OutputAllNewsView(APIView):
         news = Posts.objects.filter(draft=False)
         serializer = OutputAllNews(news, many=True)
         return Response(serializer.data)
+
+
+class AccountDetail(APIView):
+
+    def get_user(self, pk):
+        try:
+            return User.objects.get(username=pk)
+        except User.DoesNotExist:
+            raise Http404
+
+    def get(self, request, pk, format=None): #отдавать не все параметры
+        user = self.get_user(pk)
+        account = Account.objects.get(user=user)
+        serializer = AccountSerializer(account)
+        return Response(serializer.data)
+
+    def put(self, request, pk, format=None):
+        user = request.user
+        account = Account.objects.get(user=user)
+        serializer = AccountSerializer(account, data=request.data)
+        if serializer.is_valid() and user.is_authenticated and user.username == pk:
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk, format=None):
+        user = request.user
+        if user.is_authenticated and user.username == pk:
+                user.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
 
 class PostView(APIView):
     """ Вывод страницы записи"""
@@ -61,12 +89,12 @@ class PostView(APIView):
     def delete(self, request, pk, format=None):
         posts = Posts.objects.get(id=pk)
         account = Account.objects.get(user=request.user)
-        if account.is_developer and posts.author == request.user:
+        if account.is_developer and posts.user == request.user:
             posts.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-class CommentCreateView(APIView):
+class CommentNewsCreateView(APIView):
     """Добавление комментария на страницу записи"""
     """
     Help text: комментарий оставляет авторизованный пользователь User.
@@ -84,6 +112,56 @@ class CommentCreateView(APIView):
             comment.save(user=user, page=posts)
             return Response(comment.data)
         return Response(comment.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def put(self, request, pk):
+        comment = Comments_Post.objects.get(id=pk)
+        serializer = CommentsNewsSerializer(comment, data=request.data)
+        user = request.user
+        if serializer.is_valid() and user.is_authenticated and comment.user == user:
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk, format=None):
+        comment = Comments_Post.objects.get(id=pk)
+        if comment.user == request.user:
+            comment.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+class CommentGameCreateView(APIView):
+    """Добавление комментария на страницу игры"""
+    """
+    Help text: комментарий оставляет авторизованный пользователь User.
+    Оставляет его на странице game. Страницу мы ищем по url страницы, который нам отправит фронт
+    Проверяем зарегистрирован ли пользователь и валидный ли комментарий и сохраняем его в базу с соответствующими параметрами
+    
+    Важно! Если зайти в сериализатор комментария, в него не отправляется Page, 
+    т.к. страница у нас определяется по url (то что вверху описано).
+    """
+    def post(self, request):
+        user = request.user
+        comment = CommentsGameSerializer(data=request.data)
+        game = Game.objects.get(url=request.data["page"])    # Ищем пост, к которому был оставлен коммент. По урлу.
+        if comment.is_valid() and user.is_authenticated:
+            comment.save(user=user, page=game)
+            return Response(comment.data)
+        return Response(comment.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def put(self, request, pk):
+        comment = Comments_Game.objects.get(id=pk)
+        serializer = CommentsGameSerializer(comment, data=request.data)
+        user = request.user
+        if serializer.is_valid() and user.is_authenticated and comment.author == user:
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk, format=None):
+        comment = Comments_Game.objects.get(id=pk)
+        if comment.author == request.user:
+            comment.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class GameDetailView(RetrieveUpdateDestroyAPIView):
@@ -103,8 +181,7 @@ class GameDetail(APIView):
     def post(self, request, format=None):
         user = request.user
         serializer = GameSerializer(data=request.data)
-        if serializer.is_valid():
-            if user.is_authenticated:
+        if serializer.is_valid() and user.is_authenticated:
                 account = Account.objects.get(user=user)
                 if account.is_developer:
                     serializer.save(author=account)
@@ -120,11 +197,9 @@ class GameDetail(APIView):
         game = self.get_game(pk)
         serializer = GameSerializer(game, data=request.data)
         user = request.user
-        if serializer.is_valid():
-            if user.is_authenticated:
+        if serializer.is_valid() and user.is_authenticated:
                 account = Account.objects.get(user=user)
-                if account.is_developer:
-                    if game.author == account:
+                if account.is_developer and game.author == account:
                         serializer.save()
                         return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -132,8 +207,7 @@ class GameDetail(APIView):
     def delete(self, request, pk, format=None):
         game = self.get_game(pk)
         account = Account.objects.get(user=request.user)
-        if account.is_developer:
-            if game.author == account:
+        if account.is_developer and game.author == account:
                 game.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -150,8 +224,7 @@ class GameRatingDetail(APIView):
         user = request.user
         mark = request.POST.get('mark')
         rating_serializer = RatingSerializer(data={'author': user.id, 'mark': mark, 'game': pk})
-        if rating_serializer.is_valid():
-            if user.is_authenticated:
+        if rating_serializer.is_valid() and user.is_authenticated:
                 user_ratings = Rating.objects.filter(author=user.id)
                 flag = True
                 for rating in user_ratings:
@@ -221,6 +294,10 @@ class BuyGameDetail(APIView):
                 if account not in game.players.all():
                     game.players.add(account)
                     game.who_added_to_wishlist.remove(account)
+                    price = game.price * game.sale_percent / 100
+                    serializer = OrderSerializer(data={'user': user.id, 'game': pk, 'price': price, 'date': date.today()})
+                    if serializer.is_valid():
+                        serializer.save()
                     return Response({"message": "success"}, status=status.HTTP_200_OK)
                 return Response({"message": "bought"})
             except Account.DoesNotExist:
@@ -311,10 +388,17 @@ class AssessPostDetail(APIView):
 
 class OutputLibrary(ListAPIView):
     """Вывод библиотеки игр пользователя"""
+
     def get(self, request, pk):
-        games = Game.objects.filter(players=pk)
-        serializer = GameSerializer(games, many=True)
-        return Response(serializer.data)
+        user = request.user
+        account = Account.objects.get(user=user)
+        if user.username == pk:
+            games = Game.objects.filter(players=account)
+            serializer = GameSerializer(games, many=True)
+            return Response(serializer.data)
+        #else:
+        #выводить другие данные игры
+        return Response(status=status.HTTP_400_BAD_REQUEST)
 
 
 class DownloadGame(ListAPIView):
@@ -326,8 +410,29 @@ class DownloadGame(ListAPIView):
             try:
                 account = Account.objects.get(user=user)
                 if account in game.players.all():
-                    # реализация скачивания
-                    return Response({"message": "success"})
+                    if game.file:
+                        response = HttpResponse(game.file)
+                        #надо заставлять разрабов загружать только зипы? сохранять имя и тип файла?
+                        response['Content-Disposition'] = 'attachment; filename=' + game.title + '.zip'
+                        return response
             except Account.DoesNotExist:
                 return Response({"message": "fail"})
         return Response({"message": "fail"})
+
+
+class GameCategoryDetail(ListAPIView):
+    """Вывод игр соответствующей категории"""
+    def get(self, request, pk):
+        category = Category.objects.get(name=pk) #при прикручивании заюзать скрипт перевода в транслит(?)
+        games = Game.objects.filter(categories=category)
+        serializer = GameSerializer(games, many=True)
+        return Response(serializer.data)
+
+
+class FaqDetail(ListAPIView):
+    """Вывод вопросов и ответов"""
+
+    def get(self, request):
+        faq = FAQ.objects.all()
+        serializer = FaqSerializer(faq, many=True)
+        return Response(serializer.data)
